// src/app/models/file.ts
export interface fileData  {
    groupName: string;
    groupId: string;
    memberId: string;
    lastName: string;
    firstName: string;
    gender: string;
    relCode: string;
    perCode: string;
    verNo: string;
    dob: string;
    createdDate: string;
    track1: string;
    track2: string;
    updMember: string;
    effDate: string;
    fullName: string;
    fileName: string;
    createdUser: string;
    expTime: string;
}